refactor to make package
