+++ pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = [
+    "setuptools==69.5.1",
+    "torch==2.0.1+cu118",
+]
+build-backend = "setuptools.build_meta"
