+++ requirements.txt
@@ -1,0 +1,5 @@
+torch>=1.7.1
+torchvision>=0.8.2
+scikit-image
+h5py
+# ninja
