+++ setup.cfg
@@ -1,0 +1,39 @@
+[metadata]
+name = syndiff
+version = attr: syndiff.__version__
+requires-python = ">=3.9"
+
+[options]
+package_dir=
+    =src
+packages=find:
+install_requires =
+    torch>=1.7.1
+    torchvision>=0.8.2
+    scikit-image
+    h5py
+
+[options.packages.find]
+where=src
+
+[options.extras_require]
+io =
+    tensorflow
+
+[options.entry_points]
+console_scripts =
+    syndiff-train = syndiff.train:main
+    syndiff-test = syndiff.test:main
+
+[mypy]
+disallow_untyped_defs = false
+warn_unused_configs = true
+warn_redundant_casts = true
+warn_unused_ignores = true
+warn_return_any = false
+strict_equality = true
+show_column_numbers = true
+show_error_codes = true
+
+[mypy-scipy.*,setuptools,sitk_cli,itk,xlrd,cv2,torchvision.*,h5py]
+ignore_missing_imports = true
