+++ src/syndiff/__init__.py
@@ -1,0 +1,1 @@
+from .version import __version__
