+++ src/syndiff/backbones/discriminator.py
