+++ src/syndiff/backbones/generator_resnet.py
