+++ src/syndiff/utils/op/__init__.py
