+++ src/syndiff/utils/utils.py
@@ -1,6 +1,6 @@
 
 import torch
-import tensorflow as tf
+import tensorflow as tf # type: ignore [import-not-found]
 import os
 import logging
